{"schema":"decodex/commit/1","summary":"Split ELF MCP server tool routers","authority":"manual"}
