+++ apps/elf-mcp/src/app/server/schemas.rs
@@ -1,12 +1,12 @@
-#[path = "schemas/admin.rs"] mod admin;
-#[path = "schemas/docs.rs"] mod docs;
-#[path = "schemas/events.rs"] mod events;
-#[path = "schemas/graph.rs"] mod graph;
-#[path = "schemas/memory.rs"] mod memory;
-#[path = "schemas/notes.rs"] mod notes;
-#[path = "schemas/search.rs"] mod search;
-#[path = "schemas/sharing.rs"] mod sharing;
-#[path = "schemas/work_journal.rs"] mod work_journal;
+mod admin;
+mod docs;
+mod events;
+mod graph;
+mod memory;
+mod notes;
+mod search;
+mod sharing;
+mod work_journal;
 
 pub(in crate::app::server) use self::{
 	admin::{
