+++ apps/elf-mcp/src/app/server/schemas/admin.rs
