+++ apps/elf-mcp/src/app/server/schemas/docs.rs
