+++ apps/elf-mcp/src/app/server/schemas/events.rs
