+++ apps/elf-mcp/src/app/server/schemas/graph.rs
