+++ apps/elf-mcp/src/app/server/schemas/memory.rs
