+++ apps/elf-mcp/src/app/server/schemas/notes.rs
