+++ apps/elf-mcp/src/app/server/schemas/search.rs
