+++ apps/elf-mcp/src/app/server/schemas/work_journal.rs
