+++ apps/elf-mcp/src/app/server/state.rs
