+++ apps/elf-mcp/src/app/server/support.rs
