+++ apps/elf-mcp/src/app/server/tests.rs
