+++ apps/elf-mcp/src/app/server/tools.rs
@@ -1,0 +1,2 @@
+mod admin;
+mod docs;
